Rename a header
