+++ libgadget/omega_nu_single.hpp
